p4: merge icmp type + code fields to work around xilinx parser limitation
It appears that xilinx RTL does not support using verify() in a
parser state which references a header that was extract()'d in a
previous state.

Merge the icmp (v4/v6) type and code fields into a single value to
allow the state transition to effectively check both the type and code
at the same time rather than trying to defer the verify() of the
code field into the next state.

This should avoid the xilinx parser limitation by removing the "late"
calls to verify().