p4: load balance on udp dst ports 16384 (0x4000) to 32767 (0x7FFF)
This will allow UDP senders to select any valid UDP destination
port number in the range, and use that dst port for all packets in a
given event.  Choosing a random (or at least spanning) set of udp dst
ports for each subsequent event will ensure that any intermediate
network equipment will see sufficient entropy in the IP 5-tuple to
spread the stream of events uniformly across ECMP hops and/or members
in a LAG.

Note that this range includes the previously specified destination
port of 19522 (0x4c42) (LB).